Changed Thread::getProcessId() return type from int to size_t to avoid 64bit warnings.
git-svn-id: http://www.openscenegraph.org/svn/osg/OpenThreads/trunk@5495 16af8721-9629-0410-8352-f15c8da7e697

+++ pthread_src/PThread.c++
@@ -484,13 +484,13 @@ int Thread::getThreadId() {
 //
 // Use: public
 //
-int Thread::getProcessId() {
+size_t Thread::getProcessId() {
 
     PThreadPrivateData *pd = static_cast<PThreadPrivateData *> (_prvData);
 
-    if(pd->idSet == false) return (unsigned int) pthread_self();
+    if(pd->idSet == false) return (size_t)(pthread_self());
 
-    return (int)(pd->tid);
+    return (size_t)(pd->tid);
 }
 
 //-----------------------------------------------------------------------------
